Crate google_cloud_secretmanager_v1

Google Cloud Client Libraries for Rust - Secret Manager API

This crate contains traits, types, and functions to interact with Secret Manager API Most applications will use the structs defined in the client module.

The client library types and functions are stable and not expected to change. Please note that Google Cloud services do change from time to time. The client libraries are designed to preserve backwards compatibility when the service changes in compatible ways. For example, adding RPCs, or fields to messages should not introduce breaking changes to the client libraries.

Available Clients

SecretManagerService

Modules

builder

Request builders.

client

Concrete implementations of this client library traits.

model

The messages and enums that are part of this client library.

stub

Traits to mock the clients in this library.

Structs

Error

The core error returned by all client libraries.

Type Aliases

Result

An alias of std::result::Result where the error is always Error.