google_cloud_secretmanager_v1

Module model

The messages and enums that are part of this client library.

Modules

replication

Defines additional types related to Replication.

replication_status

Defines additional types related to ReplicationStatus.

secret

Defines additional types related to Secret.

secret_version

Defines additional types related to SecretVersion.

Structs

AccessSecretVersionRequest

Request message for SecretManagerService.AccessSecretVersion.

AccessSecretVersionResponse

Response message for SecretManagerService.AccessSecretVersion.

AddSecretVersionRequest

Request message for SecretManagerService.AddSecretVersion.

CreateSecretRequest

Request message for SecretManagerService.CreateSecret.

CustomerManagedEncryption

Configuration for encrypting secret payloads using customer-managed encryption keys (CMEK).

CustomerManagedEncryptionStatus

Describes the status of customer-managed encryption.

DeleteSecretRequest

Request message for SecretManagerService.DeleteSecret.

DestroySecretVersionRequest

Request message for SecretManagerService.DestroySecretVersion.

DisableSecretVersionRequest

Request message for SecretManagerService.DisableSecretVersion.

EnableSecretVersionRequest

Request message for SecretManagerService.EnableSecretVersion.

GetSecretRequest

Request message for SecretManagerService.GetSecret.

GetSecretVersionRequest

Request message for SecretManagerService.GetSecretVersion.

ListSecretVersionsRequest

Request message for SecretManagerService.ListSecretVersions.

ListSecretVersionsResponse

Response message for SecretManagerService.ListSecretVersions.

ListSecretsRequest

Request message for SecretManagerService.ListSecrets.

ListSecretsResponse

Response message for SecretManagerService.ListSecrets.

Replication

A policy that defines the replication and encryption configuration of data.

ReplicationStatus

The replication status of a SecretVersion.

Rotation

The rotation time and period for a Secret. At next_rotation_time, Secret Manager will send a Pub/Sub notification to the topics configured on the Secret. Secret.topics must be set to configure rotation.

Secret

A Secret is a logical secret whose value and versions can be accessed.

SecretPayload

A secret payload resource in the Secret Manager API. This contains the sensitive secret payload that is associated with a SecretVersion.

SecretVersion

A secret version resource in the Secret Manager API.

Topic

A Pub/Sub topic which Secret Manager will publish to when control plane events occur on this secret.

UpdateSecretRequest

Request message for SecretManagerService.UpdateSecret.