google_cloud_secretmanager_v1::model

Module replication

Defines additional types related to Replication.

Modules

user_managed

Defines additional types related to UserManaged.

Structs

Automatic

A replication policy that replicates the Secret payload without any restrictions.

UserManaged

A replication policy that replicates the Secret payload into the locations specified in Replication.UserManaged.replicas

Enums

Replication

The replication policy for this secret.