google_cloud_secretmanager_v1::model::secret_version

Enum State

#[non_exhaustive]
pub enum State {
    Unspecified,
    Enabled,
    Disabled,
    Destroyed,
    UnknownValue(UnknownValue),
}

The state of a SecretVersion, indicating if it can be accessed.

Working with unknown values

This enum is defined as #[non_exhaustive] because Google Cloud may add additional enum variants at any time. Adding new variants is not considered a breaking change. Applications should write their code in anticipation of:

New values appearing in future releases of the client library, and

New values received dynamically, without application changes.

Please consult the Working with enums section in the user guide for some guidelines.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Unspecified

Not specified. This value is unused and invalid.

Enabled

The SecretVersion may be accessed.

Disabled

The SecretVersion may not be accessed, but the secret data is still available and can be placed back into the ENABLED state.

Destroyed

The SecretVersion is destroyed and the secret data is no longer stored. A version may not leave this state once entered.

UnknownValue(UnknownValue)

If set, the enum was initialized with an unknown value.

Applications can examine the value using State::value or State::name.

Implementations

impl State

pub fn value(&self) -> Option<i32>

Gets the enum value.

Returns None if the enum contains an unknown value deserialized from the string representation of enums.

pub fn name(&self) -> Option<&str>

Gets the enum value as a string.

Returns None if the enum contains an unknown value deserialized from the integer representation of enums.