google_cloud_secretmanager_v1::stub

Trait SecretManagerService

pub trait SecretManagerService:
    Debug
    + Send
    + Sync {

    // Provided methods
    fn list_secrets(
        &self,
        _req: ListSecretsRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<ListSecretsResponse>>> + Send { ... }
    fn create_secret(
        &self,
        _req: CreateSecretRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<Secret>>> + Send { ... }
    fn add_secret_version(
        &self,
        _req: AddSecretVersionRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send { ... }
    fn get_secret(
        &self,
        _req: GetSecretRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<Secret>>> + Send { ... }
    fn update_secret(
        &self,
        _req: UpdateSecretRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<Secret>>> + Send { ... }
    fn delete_secret(
        &self,
        _req: DeleteSecretRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<()>>> + Send { ... }
    fn list_secret_versions(
        &self,
        _req: ListSecretVersionsRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<ListSecretVersionsResponse>>> + Send { ... }
    fn get_secret_version(
        &self,
        _req: GetSecretVersionRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send { ... }
    fn access_secret_version(
        &self,
        _req: AccessSecretVersionRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<AccessSecretVersionResponse>>> + Send { ... }
    fn disable_secret_version(
        &self,
        _req: DisableSecretVersionRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send { ... }
    fn enable_secret_version(
        &self,
        _req: EnableSecretVersionRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send { ... }
    fn destroy_secret_version(
        &self,
        _req: DestroySecretVersionRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send { ... }
    fn set_iam_policy(
        &self,
        _req: SetIamPolicyRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<Policy>>> + Send { ... }
    fn get_iam_policy(
        &self,
        _req: GetIamPolicyRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<Policy>>> + Send { ... }
    fn test_iam_permissions(
        &self,
        _req: TestIamPermissionsRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<TestIamPermissionsResponse>>> + Send { ... }
    fn list_locations(
        &self,
        _req: ListLocationsRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<ListLocationsResponse>>> + Send { ... }
    fn get_location(
        &self,
        _req: GetLocationRequest,
        _options: RequestOptions,
    ) -> impl Future<Output = Result<Response<Location>>> + Send { ... }

}

Defines the trait used to implement super::client::SecretManagerService.

Application developers may need to implement this trait to mock client::SecretManagerService. In other use-cases, application developers only use client::SecretManagerService and need not be concerned with this trait or its implementations.

Services gain new RPCs routinely. Consequently, this trait gains new methods too. To avoid breaking applications the trait provides a default implementation of each method. Most of these implementations just return an error.

Provided Methods

fn list_secrets( &self, _req: ListSecretsRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<ListSecretsResponse>>> + Send

Implements super::client::SecretManagerService::list_secrets.

fn create_secret( &self, _req: CreateSecretRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<Secret>>> + Send

Implements super::client::SecretManagerService::create_secret.

fn add_secret_version( &self, _req: AddSecretVersionRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send

Implements super::client::SecretManagerService::add_secret_version.

fn get_secret( &self, _req: GetSecretRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<Secret>>> + Send

Implements super::client::SecretManagerService::get_secret.

fn update_secret( &self, _req: UpdateSecretRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<Secret>>> + Send

Implements super::client::SecretManagerService::update_secret.

fn delete_secret( &self, _req: DeleteSecretRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<()>>> + Send

Implements super::client::SecretManagerService::delete_secret.

fn list_secret_versions( &self, _req: ListSecretVersionsRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<ListSecretVersionsResponse>>> + Send

Implements super::client::SecretManagerService::list_secret_versions.

fn get_secret_version( &self, _req: GetSecretVersionRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send

Implements super::client::SecretManagerService::get_secret_version.

fn access_secret_version( &self, _req: AccessSecretVersionRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<AccessSecretVersionResponse>>> + Send

Implements super::client::SecretManagerService::access_secret_version.

fn disable_secret_version( &self, _req: DisableSecretVersionRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send

Implements super::client::SecretManagerService::disable_secret_version.

fn enable_secret_version( &self, _req: EnableSecretVersionRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send

Implements super::client::SecretManagerService::enable_secret_version.

fn destroy_secret_version( &self, _req: DestroySecretVersionRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<SecretVersion>>> + Send

Implements super::client::SecretManagerService::destroy_secret_version.

fn set_iam_policy( &self, _req: SetIamPolicyRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<Policy>>> + Send

Implements super::client::SecretManagerService::set_iam_policy.

fn get_iam_policy( &self, _req: GetIamPolicyRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<Policy>>> + Send

Implements super::client::SecretManagerService::get_iam_policy.

fn test_iam_permissions( &self, _req: TestIamPermissionsRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<TestIamPermissionsResponse>>> + Send

Implements super::client::SecretManagerService::test_iam_permissions.

fn list_locations( &self, _req: ListLocationsRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<ListLocationsResponse>>> + Send

Implements super::client::SecretManagerService::list_locations.

fn get_location( &self, _req: GetLocationRequest, _options: RequestOptions, ) -> impl Future<Output = Result<Response<Location>>> + Send

Implements super::client::SecretManagerService::get_location.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.